requirements.txt
bcrypt==3.1.7
cffi==1.14.0
click==7.1.1
dnspython==1.16.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-PyMongo==2.3.0
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.1.1
pycparser==2.20
pymongo==3.10.1
six==1.14.0
Werkzeug==1.0.1
WTForms==2.2.1